Add KCP APIs to scheme, enforce kind image version
While testing out the `controller-runtime-example` Makefile,
specifically running `make test-e2e` I noticed that the build was
failing locally and the controller pod was crashing.

After a bit of debugging, it seems that the scheme was missing KCP's
APIExport Kind, which caused the `error looking up virtual workspace
URL` error, and then the panic for the controller.

In addition, this PR enforces a known working version of Kubernetes when
testing locally, and it fixes an issue on Mac where the ~ sign (i.e. HOME) wasn't
working properly.

Signed-off-by: Vince Prignano <vince@prigna.com>